test: restructure file formatting
